chore: remove the build script, use `?`, fix typo, update serde
